Update FtpSharp api to v2.2.0
