+++ SignalAnalysis/FftSharp/Windows/Bartlett.cs
@@ -1,31 +1,30 @@
 ﻿using System;
 
-namespace FftSharp.Windows
+namespace FftSharp.Windows;
+
+public class Bartlett : Window, IWindow
 {
-    public class Bartlett : Window, IWindow
+    public override string Name => "Bartlett–Hann";
+    public override string Description =>
+        "The Bartlett–Hann window is triangular in shape (a 2nd order B-spline) which is effectively the " +
+        "convolution of two half-sized rectangular windows.";
+
+    public override bool IsSymmetric => true;
+
+    public override double[] Create(int size, bool normalize = false)
     {
-        public override string Name => "Bartlett–Hann";
-        public override string Description =>
-            "The Bartlett–Hann window is triangular in shape (a 2nd order B-spline) which is effectively the " +
-            "convolution of two half-sized rectangular windows.";
+        double[] window = new double[size];
 
-        public override bool IsSymmetric => true;
+        bool isOddSize = size % 2 == 1;
 
-        public override double[] Create(int size, bool normalize = false)
-        {
-            double[] window = new double[size];
+        double halfSize = isOddSize ? size / 2 : (size - 1) / 2.0;
 
-            bool isOddSize = size % 2 == 1;
+        for (int i = 0; i < size; i++)
+            window[i] = 1 - Math.Abs((double)(i - halfSize) / halfSize);
 
-            double halfSize = isOddSize ? size / 2 : (size - 1) / 2.0;
+        if (normalize)
+            NormalizeInPlace(window);
 
-            for (int i = 0; i < size; i++)
-                window[i] = 1 - Math.Abs((double)(i - halfSize) / halfSize);
-
-            if (normalize)
-                NormalizeInPlace(window);
-
-            return window;
-        }
+        return window;
     }
 }
