+++ SignalAnalysis/FftSharp/Windows/Cosine.cs
@@ -1,25 +1,24 @@
 ﻿using System;
 using System.Linq;
 
-namespace FftSharp.Windows
+namespace FftSharp.Windows;
+
+public class Cosine : Window, IWindow
 {
-    public class Cosine : Window, IWindow
+    public override string Name => "Cosine";
+    public override string Description =>
+        "This window is simply a cosine function. It reaches zero on both sides and is similar to " +
+        "Blackman, Hamming, Hanning, and flat top windows, but probably should not be used in practice.";
+
+    public override bool IsSymmetric => true;
+
+    public override double[] Create(int size, bool normalize = false)
     {
-        public override string Name => "Cosine";
-        public override string Description =>
-            "This window is simply a cosine function. It reaches zero on both sides and is similar to " +
-            "Blackman, Hamming, Hanning, and flat top windows, but probably should not be used in practice.";
+        double[] window = Enumerable.Range(0, size).Select(x => Math.Sin(Math.PI / (size) * (x + .5))).ToArray();
 
-        public override bool IsSymmetric => true;
+        if (normalize)
+            NormalizeInPlace(window);
 
-        public override double[] Create(int size, bool normalize = false)
-        {
-            double[] window = Enumerable.Range(0, size).Select(x => Math.Sin(Math.PI / (size) * (x + .5))).ToArray();
-
-            if (normalize)
-                NormalizeInPlace(window);
-
-            return window;
-        }
+        return window;
     }
 }
