+++ SignalAnalysis/FftSharp/Windows/Rectangular.cs
@@ -1,26 +1,25 @@
-﻿namespace FftSharp.Windows
+﻿namespace FftSharp.Windows;
+
+public class Rectangular : Window, IWindow
 {
-    public class Rectangular : Window, IWindow
+    public override string Name => "Rectangular";
+    public override string Description =>
+        "The rectangular window (sometimes known as the boxcar or Dirichlet window) is the simplest window, " +
+        "equivalent to replacing all but N values of a data sequence by zeros, making it appear as though " +
+        "the waveform suddenly turns on and off. This window preserves transients at the start and end of the signal.";
+
+    public override bool IsSymmetric => true;
+
+    public override double[] Create(int size, bool normalize = false)
     {
-        public override string Name => "Rectangular";
-        public override string Description =>
-            "The rectangular window (sometimes known as the boxcar or Dirichlet window) is the simplest window, " +
-            "equivalent to replacing all but N values of a data sequence by zeros, making it appear as though " +
-            "the waveform suddenly turns on and off. This window preserves transients at the start and end of the signal.";
+        double[] window = new double[size];
 
-        public override bool IsSymmetric => true;
+        for (int i = 0; i < size; i++)
+            window[i] = 1;
 
-        public override double[] Create(int size, bool normalize = false)
-        {
-            double[] window = new double[size];
+        if (normalize)
+            NormalizeInPlace(window);
 
-            for (int i = 0; i < size; i++)
-                window[i] = 1;
-
-            if (normalize)
-                NormalizeInPlace(window);
-
-            return window;
-        }
+        return window;
     }
 }
